:truck: Moved types about `window.scrapbox` to userscript.ts
